+++ internal/cmd/root.go
@@ -105,8 +105,8 @@ func initRootFlags(rootCmd *cobra.Command) {
 	rootCmd.Flags().StringArray("exclude", []string{}, "run all but excluded command(s) described in cmd as map")
 	rootCmd.Flags().Bool("init", false, "create a new lets.yaml in the current folder")
 	rootCmd.Flags().Bool("no-depends", false, "skip 'depends' for running command")
-	rootCmd.Flags().CountP("debug", "d", "show debug logs (or use LETS_DEBUG=1). If used multiple times, shows more verbose logs") //nolint:lll
+	rootCmd.Flags().CountP("debug", "d", "show debug logs; repeat for verbose output")
 	rootCmd.Flags().StringP("config", "c", "", "config file (default is lets.yaml)")
 	rootCmd.Flags().Bool("all", false, "show all commands (including the ones with _)")
-	rootCmd.Flags().Bool("no-cache", false, "re-download remote config instead of using cached version")
+	rootCmd.Flags().Bool("no-cache", false, "re-download remote config and mixins instead of using cached versions")
 }